extract marching cube from metaball
now a class by itself
